Allow for rolling own binary
