+++ example/pkgname/user.go
@@ -1,3 +1,3 @@
 package differentpkg
 
-//go:generate go run ../../dataloaden.go ../../template.go -keys string github.com/vektah/dataloaden/example.User
+//go:generate go run ../../dataloaden.go -keys string github.com/vektah/dataloaden/example.User
